requirements.txt
asgiref==3.10.0
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
Django==4.2.25
django-cors-headers==4.9.0
django-stubs==5.1.3
django-stubs-ext==5.1.3
djangorestframework==3.16.1
djangorestframework-stubs==3.15.3
djangorestframework_simplejwt==5.5.1
exceptiongroup==1.3.1
idna==3.11
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-django==4.11.1
requests==2.32.5
sqlparse==0.5.3
tomli==2.3.0
types-PyYAML==6.0.12.20250915
types-requests==2.32.4.20250913
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
